Moved db initialization to the app layer
